+++ lostandfound/src/main/java/com/example/lostandfound/repository/ReportRepository.java
@@ -1,10 +1,11 @@
 package com.example.lostandfound.repository;
 
-import com.example.lostandfound.model.Report;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.example.lostandfound.model.Report;
+
 @Repository
-public interface ReportRepository extends JpaRepository<Report, Long> {
+public interface ReportRepository extends JpaRepository<Report, Integer> {
     // Additional custom queries can be added here
 }
